+++ projects/ion/src/lib/core/types/tag.ts
@@ -1,0 +1,9 @@
+import { TagStatus } from './status';
+
+export interface IonTagProps {
+  outline?: boolean;
+  status?: TagStatus;
+  color?: string;
+  label: string;
+  icon?: string;
+}
