+++ projects/ion/src/lib/tag/tag.component.spec.ts
@@ -1,8 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { render, screen } from '@testing-library/angular';
-import { TagStatus } from './../core/types';
-import { IonTagProps, IonTagComponent } from './tag.component';
 import { IonIconModule } from '../icon/icon.module';
+import { IonTagProps, TagStatus } from './../core/types';
+import { IonTagComponent } from './tag.component';
 
 const defaultValue: IonTagProps = {
   label: 'Default Message',
